script: Support LSP signature help (#642)
* script: Minor expr query re-order

* script: Support filtered finds

* script: Refactor script find routine

* lsp: Signature-help dummy impl

* lsp: Initial expr lookup in signature-help

* lsp: Find exprs with a signature

* lsp: Get signature for sig help

* script: Add script arg write util

* script: Fix incorrect arg count type in read.c

* lsp: Initial signature help impl

* lsp: Add label to signature text

* script: Add script_expr_arg_index util

* lsp: Report active parameter

* script: Fix implicit int to u32 conversion

+++ libs/script/src/read.c
@@ -759,9 +759,9 @@ read_emit_unreachable(ScriptReadContext* ctx, const ScriptExpr exprs[], const u3
       const ScriptPos  unreachableStart = expr_range(ctx->doc, exprs[i + 1]).start;
       const ScriptPos  unreachableEnd   = expr_range(ctx->doc, exprs[exprCount - 1]).end;
       const ScriptDiag unreachableDiag  = {
-          .severity = ScriptDiagSeverity_Warning,
-          .kind     = ScriptDiag_ExprUnreachable,
-          .range    = script_range(unreachableStart, unreachableEnd),
+           .severity = ScriptDiagSeverity_Warning,
+           .kind     = ScriptDiag_ExprUnreachable,
+           .range    = script_range(unreachableStart, unreachableEnd),
       };
       script_diag_push(ctx->diags, &unreachableDiag);
       break;
@@ -1096,7 +1096,7 @@ read_expr_call(ScriptReadContext* ctx, const StringHash id, const ScriptRange id
 
   const ScriptBuiltinFunc* builtin = script_builtin_func_lookup(id);
   if (builtin) {
-    const u16 expectedArgCount = script_sig_arg_count(builtin->sig);
+    const u8 expectedArgCount = script_sig_arg_count(builtin->sig);
     if (UNLIKELY(expectedArgCount != argCount)) {
       read_emit_err(ctx, ScriptDiag_IncorrectArgCountForBuiltinFunc, callRange);
 
@@ -1108,7 +1108,7 @@ read_expr_call(ScriptReadContext* ctx, const StringHash id, const ScriptRange id
        * sure the program is well formed.
        */
       const ScriptPos lastPos = callRange.end - 1;
-      for (u16 i = (u16)argCount; i < expectedArgCount; ++i) {
+      for (u8 i = (u8)argCount; i < expectedArgCount; ++i) {
         args[i] = script_add_value(ctx->doc, script_range(lastPos, lastPos), script_null());
       }
       ctx->flags |= ScriptReadFlags_ProgramInvalid;
@@ -1736,13 +1736,13 @@ ScriptExpr script_read(
 
   ScriptScope       scopeRoot = {0};
   ScriptReadContext ctx       = {
-      .doc        = doc,
-      .binder     = binder,
-      .diags      = diags,
-      .syms       = syms,
-      .input      = src,
-      .inputTotal = src,
-      .scopeRoot  = &scopeRoot,
+            .doc        = doc,
+            .binder     = binder,
+            .diags      = diags,
+            .syms       = syms,
+            .input      = src,
+            .inputTotal = src,
+            .scopeRoot  = &scopeRoot,
   };
   read_var_free_all(&ctx);
